Update 190515_cutadapt_pilot_GBS
Changed algorithm being used (aln -> mem) and changed path of fastq file (and ran the script within the directory for the sample rather than ~/pilot_GBS. When I first ran this code, bwa would successfully create the index, but it would not write to the SAM file (or SAI in the case of the aln algorithm). Changing the path to the fastq file (and changing directories to make up for the new path name) fixed this problem.