Initial commit
This file gets SNP stats for samples from pilot study (number of SNPs per scaffold & min, max, mean, median distance between SNPs).
Purpose was to have it ready in time for field day. Implementation could be improved, but this was the quick and dirty way to just have it done.