Updated directory structure
Additional GBS files were released by UMGC, so this code was updated to make symlinks to those files. They still need to be concatenated with the  samples that were released on August 15. Basically, the updated code makes links with the generalized name "Sample_[number]_2.fq.gz. They still need to be concatenated so that the analysis can be done on the full data for each file.