Update to file pca_from_VCF
This code creates a PCA plot for the first 50 samples (and only for Scaffold 48). The reason for doing the subset is that it would be less computationally intensive and I wanted at least some visualization of my data... But it still took a very long time to calculate the principal components.