Initial commit
In this file, I resolve some of the problems with missing data that plagued the main GBS analysis so far. I achieved better results by using a lower depth threshold (DP > 5 vs. DP > 10) to allow for more SNPs. The result is fewer missing data (NAs).  With DP > 10, only 28 SNPs had fewer than 15% missing data (# of samples where the SNP was present). With DP > 5, there are 341 SNPs with less than 15% missing data. The aim was to have ~2000 SNPs... which we do have, but some of them have higher rates of missing data.

This is an update to: 191010_main_GBS_snp_calling_troubleshooting.R. Only 200 samples were included. The next step is to utilize all 1054 samples.