Initial commit
This file still works on the 200-member subset of the main_GBS panel, but it uses a mapping quality threshold of 40 (vs. 60) and a depth of 5 (vs 10). The result is ~2700 SNPs with 30 or fewer NAs (out of 200; 15%). This was the last of the "troubleshooting" files that gave me the confidence that it was appropriate to move forward with the analysis of the full main_GBS set (1054 samples).