Initial commit
These two text files contain mostly R code (aside from command-line code at the beginning to generate necessary PLINK files). The purpose is to do some sleuthing as to why the PLINK PCA looks different than the PCA from the R package poppr. The results suggest that the presence of _Z. latifolia_ and other pilot samples do not make a difference in the % variation explained. Apart from that, it was needed to run the code without the pilot samples anyway since we decided to exclude them due to differences in sequencing depth. The next suspect to test in reconciling differences in PCA plots is the number of SNPs used/missing data (maybe attributed to _Z. aquatica_ samples?