Update
The method to calculate percent variation was correct before.. The changes I made earlier today gave an answer that I thought was more accurate. The values were almost spot on to the expectation for % variation (based on PLINK plots) but the math I used to arrive at the answer was wrong. The changes here are to use only the top 20 eigenvalue in determining % variation rather than all of them (there are hundreds, most with very small values). I think this will solve the issue for good.