add subgroups
The code now works in such a way that the popgen statistics that I was looking for are calculated, but not on a per-population basis per se. I can't look up individual populations by Bass_Lake or any other lake/river for that matter.. Only by samples. Something likely needs to be changed.