rename object x to Fst
there was a conflict with the existing code when I updated the Fst Rdata file. I think it is because in the Rdata file from April 20, the Fst object was clearly named "Fst". But in my updated version, I think the object was just named "x" and that was causing a conflict. I think that the change I'm making here should fix that.